Feat: RP-32 소프트 딜리트 구현
