+++ src/main/java/com/cherrypick/backend/domain/deal/dto/request/DealRequestDTOs.java
@@ -20,7 +20,4 @@ public class DealRequestDTOs {
         }
     }
 
-
-
-
 }
