+++ src/main/java/com/cherrypick/backend/domain/user/repository/UserRepository.java
@@ -5,6 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
+import java.util.List;
 import java.util.Optional;
 
 public interface UserRepository extends JpaRepository<User, Long>
@@ -14,5 +15,8 @@ public interface UserRepository extends JpaRepository<User, Long>
 
     @Query(value="SELECT * FROM user WHERE nickname=:nickname;", nativeQuery = true)
     Optional<User> findUserByNickname(@Param("nickname") String nickname);
+
+    @Query(value="SELECT * FROM user WHERE status='DEACTIVATED'", nativeQuery = true)
+    List<User> findDeactivatedUsers();
 
 }
