+++ src/main/java/com/cherrypick/backend/domain/user/service/UserService.java
@@ -1,11 +1,13 @@
 package com.cherrypick.backend.domain.user.service;
 
 
+import com.cherrypick.backend.domain.deal.dto.request.DealRequestDTOs;
 import com.cherrypick.backend.domain.image.enums.ImageType;
 import com.cherrypick.backend.domain.image.service.ImageService;
 import com.cherrypick.backend.domain.user.dto.UserDetailResponseDTO;
 import com.cherrypick.backend.domain.user.dto.UserRequestDTOs;
 import com.cherrypick.backend.domain.user.dto.UserResponseDTOs;
+import com.cherrypick.backend.domain.user.dto.UserUpdateRequestDTO;
 import com.cherrypick.backend.domain.user.enums.UserStatus;
 import com.cherrypick.backend.domain.user.repository.UserRepository;
 import com.cherrypick.backend.global.exception.BaseException;
@@ -13,9 +15,11 @@ import com.cherrypick.backend.global.exception.enums.UserErrorCode;
 import com.cherrypick.backend.global.util.AuthUtil;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.time.LocalDateTime;
 import java.util.List;
 
 @Service
@@ -26,7 +30,7 @@ public class UserService {
     private final ImageService imageService;
 
     @Transactional
-    public UserDetailResponseDTO userUpdate(UserRequestDTOs.UpdateDTO dto){
+    public UserDetailResponseDTO userUpdate(UserUpdateRequestDTO dto){
 
         // 인증된 유저의 ID를 찾음, 인증되지 않았다면 오류.
         var userId = AuthUtil.getUserDetail().userId();
@@ -96,7 +100,7 @@ public class UserService {
 
     }
 
-    public UserResponseDTOs.DeleteResponseDTO softDelete() {
+    public UserResponseDTOs.DeleteResponseDTO softDelete(UserRequestDTOs.DeleteRequestDTO dto) {
 
         var userId = AuthUtil.getUserDetail().userId();
         var user = userRepository.findById(userId).orElseThrow(() -> new BaseException(UserErrorCode.USER_NOT_FOUND));
@@ -106,6 +110,14 @@ public class UserService {
         return new UserResponseDTOs.DeleteResponseDTO(deletedUserId.getUserId(), "soft delete success");
     }
 
+    /// 매일 오전 4시 정각에 실행, 성능 생각하면 리펙토링 해야함.
+    /// 하드 딜리트하는 함수
+    @Scheduled(cron = "0 0 4 * * *")
+    public void hardDelete() {
 
+        var userList = userRepository.findDeactivatedUsers();
+        userRepository.deleteAll(userList);
+        log.info("{} ::::: 총 {}명의 유저 삭제", LocalDateTime.now(), userList.size());
+    }
 
 }
